widen model tri object to int

+++ src/engine/model/skelmodel.h
@@ -51,7 +51,7 @@ struct skelmodel : animmodel
 
     struct tri
     {
-        ushort vert[3];
+        uint vert[3];
     };
 
     class blendcombo
